requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
schedule==1.2.2
aiofiles==23.2.1
jinja2==3.1.4
python-multipart==0.0.9
websockets>=10.0,<12.0
lxml==5.2.2
pyOpenSSL==24.1.0
dnspython==2.6.1
PyJWT==2.8.0
